+++ src/lib/schema-to-ts/ir.ts
@@ -36,50 +36,3 @@ export interface IRRoot {
 }
 
 export const UNKNOWN: IRNode = { kind: 'unknown' };
-
-/**
- * Structural key used only for de-duplicating union members. The canonical serializer that
- * feeds the hash is a separate, sorted representation — do not conflate them.
- */
-export function nodeKey(node: IRNode): string {
-	switch (node.kind) {
-		case 'literal':
-			return `l:${typeof node.value}:${String(node.value)}`;
-		case 'union':
-			return `u[${node.members.map(nodeKey).join(',')}]`;
-		case 'array':
-			return `a[${nodeKey(node.items)}]`;
-		case 'object':
-			return `o[${node.props
-				.map((p) => `${p.name}${p.required ? 'R' : '-'}${p.hasDefault ? 'D' : '-'}:${nodeKey(p.node)}`)
-				.join(',')}|${node.valueType ? nodeKey(node.valueType) : ''}|${node.open ? '+' : '-'}]`;
-		default:
-			return node.kind;
-	}
-}
-
-/**
- * Flattens, de-dupes, and collapses so there is exactly one IR per type.
- * @deprecated should be done by the Parser
- **/
-export function union(members: IRNode[]): IRNode {
-	const flat: IRNode[] = [];
-	const seen = new Set<string>();
-	const push = (node: IRNode): void => {
-		if (node.kind === 'union') {
-			node.members.forEach(push);
-			return;
-		}
-		const key = nodeKey(node);
-		if (seen.has(key)) return;
-		seen.add(key);
-		flat.push(node);
-	};
-	members.forEach(push);
-
-	if (flat.length === 0) return UNKNOWN;
-	if (flat.length === 1) return flat[0]!;
-	// `unknown` absorbs everything it is unioned with.
-	if (flat.some((n) => n.kind === 'unknown')) return UNKNOWN;
-	return { kind: 'union', members: flat };
-}
